Added validation code for imported submissions

+++ app/web/src/main/java/uk/ac/ebi/fg/annotare2/web/gwt/common/client/SubmissionValidationService.java
@@ -28,6 +28,6 @@ public interface SubmissionValidationService extends RemoteService {
 
     public static final String NAME = "validationService";
 
-    public ValidationResult validate(int submissionId) throws ResourceNotFoundException, NoPermissionException;
+    public ValidationResult validate(long submissionId) throws ResourceNotFoundException, NoPermissionException;
 
 }
